requirements.txt
streamlit
google-generativeai
pandas
pillow
pymupdf
openpyxl
gspread
google-auth
st-gsheets-connection
